Move TestContext reflection to helper class
